♻️ Remove old code, change package names
Replace package name to match maven artifact, remove deprecated
code, remove unused modules, add docs.

Related: #56997
