--- android/src/main/java/cz/ackee/extensions/android/CollectionsExtensions.kt
@@ -1,11 +1,0 @@
-package cz.ackee.extensions.android
-
-
-// Extensions to collections
-
-/**
- * Check if this collection is null or empty
- */
-fun Collection<*>?.isNullOrEmpty(): Boolean {
-    return this == null || this.isEmpty()
-}
