+++ build.gradle
@@ -8,9 +8,10 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath 'com.android.tools.build:gradle:4.1.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath 'io.codearte.gradle.nexus:gradle-nexus-staging-plugin:0.30.0'
+        classpath "org.jetbrains.dokka:dokka-gradle-plugin:$kotlin_version"
     }
 }
 
@@ -18,6 +19,7 @@ allprojects {
     repositories {
         jcenter()
         google()
+        mavenLocal()
     }
 }
