+++ epoxy/build.gradle
@@ -24,7 +24,7 @@ android {
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'com.airbnb.android:epoxy:4.4.2'
-    implementation "cz.ackee.extensions:anko:1.4.1"
+    implementation "io.github.ackeecz:extensions-anko:2.0.0"
     implementation "org.jetbrains.anko:anko-sdk15:0.10.8"
 }
