+++ epoxy/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="cz.ackee.extensions.epoxy">
+<manifest package="io.github.ackeecz.extensions.epoxy">
 
-    <application/>
+    <application />
 </manifest>
