+++ epoxy2/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="cz.ackee.extensions.epoxy2">
+<manifest package="io.github.ackeecz.extensions.epoxy2">
 
     <application />
 </manifest>
