--- epoxy2/src/main/java/cz/ackee/extensions/epoxy/EpoxyModelWithViewBinding.kt
@@ -1,66 +1,0 @@
-package cz.ackee.extensions.epoxy
-
-import android.view.View
-import android.view.ViewParent
-import androidx.annotation.LayoutRes
-import androidx.viewbinding.ViewBinding
-import com.airbnb.epoxy.EpoxyHolder
-import com.airbnb.epoxy.EpoxyModelWithHolder
-import java.lang.reflect.Method
-import java.lang.reflect.ParameterizedType
-import java.util.concurrent.ConcurrentHashMap
-
-/**
- * EpoxyModel class with view binding support
- *
- * Taken from https://github.com/airbnb/epoxy/blob/master/kotlinsample/src/main/java/com/airbnb/epoxy/kotlinsample/helpers/ViewBindingEpoxyModelWithHolder.kt
- */
-abstract class EpoxyModelWithViewBinding<in T : ViewBinding>(@LayoutRes val resId: Int) : EpoxyModelWithHolder<ViewBindingHolder>() {
-
-    override fun getDefaultLayout(): Int {
-        return resId
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    override fun bind(holder: ViewBindingHolder) {
-        (holder.viewBinding as T).bind()
-    }
-
-    abstract fun T.bind()
-
-    override fun createNewHolder(parent: ViewParent): ViewBindingHolder {
-        return ViewBindingHolder(this::class.java)
-    }
-}
-
-// Static cache of a method pointer for each type of item used.
-private val sBindingMethodByClass = ConcurrentHashMap<Class<*>, Method>()
-
-@Suppress("UNCHECKED_CAST")
-@Synchronized
-private fun getBindMethodFrom(javaClass: Class<*>): Method {
-    return sBindingMethodByClass.getOrPut(javaClass) {
-        val actualTypeOfThis = getSuperclassParameterizedType(javaClass)
-        val viewBindingClass = actualTypeOfThis.actualTypeArguments[0] as Class<ViewBinding>
-        viewBindingClass.getDeclaredMethod("bind", View::class.java)
-            ?: error("The binder class ${javaClass.canonicalName} should have a method bind(View)")
-    }
-}
-
-private fun getSuperclassParameterizedType(klass: Class<*>): ParameterizedType {
-    val genericSuperclass = klass.genericSuperclass
-    return (genericSuperclass as? ParameterizedType) ?: getSuperclassParameterizedType(genericSuperclass as Class<*>)
-}
-
-class ViewBindingHolder(private val epoxyModelClass: Class<*>) : EpoxyHolder() {
-    // Using reflection to get the static binding method.
-    // Lazy so it's computed only once by instance, when the 1st ViewHolder is actually created.
-    private val bindingMethod by lazy { getBindMethodFrom(epoxyModelClass) }
-
-    internal lateinit var viewBinding: ViewBinding
-
-    override fun bindView(itemView: View) {
-        // The 1st param is null because the binding method is static.
-        viewBinding = bindingMethod.invoke(null, itemView) as ViewBinding
-    }
-}
