+++ gradle/mavencentral/publish.gradle
@@ -1,5 +1,6 @@
 apply plugin: 'maven-publish'
 apply plugin: 'signing'
+apply plugin: 'org.jetbrains.dokka'
 
 task androidSourcesJar(type: Jar) {
     archiveClassifier.set('sources')
@@ -14,9 +15,16 @@ task androidSourcesJar(type: Jar) {
     }
 }
 
+task javadocJar(type: Jar, dependsOn: dokkaJavadoc) {
+    archiveClassifier.set('javadoc')
+    from dokkaJavadoc.outputDirectory
+}
+
 artifacts {
     archives androidSourcesJar
+    archives javadocJar
 }
+
 
 group = libProperties['GROUP']
 version = libProperties['VERSION_NAME']
@@ -62,6 +70,7 @@ publishing {
                 artifact("$buildDir/libs/${project.getName()}-${version}.jar")
             }
             artifact androidSourcesJar
+            artifact javadocJar
 
             // Mostly self-explanatory metadata
             pom {
