+++ lib.properties
@@ -1,5 +1,5 @@
 GROUP=io.github.ackeecz
-VERSION_NAME=1.5.0
+VERSION_NAME=2.0.0
 VERSION_CODE=1
 SITE_URL=https://github.com/AckeeCZ/kotlin-extensions
 POM_DEVELOPER_ID=ackee
@@ -7,7 +7,7 @@ POM_DEVELOPER_NAME=Ackee
 POM_DEVELOPER_EMAIL=info@ackee.cz
 POM_LICENCE_NAME=The Apache Software License, Version 2.0
 POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
-POM_SCM_CONNECTION=scm:git:github.com/ackeecz/koltin-extensions.git
+POM_SCM_CONNECTION=scm:git:github.com/ackeecz/kotlin-extensions.git
 POM_SCM_DEVELOPER_CONNECTION='scm:git:ssh://github.com/ackeecz/kotlin-extensions.git'
 POM_SCM_URL=https://github.com/ackeeCZ/kotlin-extensions/tree/main
 SONATYPE_URL=https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/
