--- recyclerview/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest package="cz.ackee.extensions.recyclerview">
-
-    <application/>
-</manifest>
