--- viewmodel/build.gradle
@@ -1,35 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-android {
-    compileSdkVersion Integer.parseInt(buildProperties["SDK_VERSION"])
-
-    defaultConfig {
-        minSdkVersion Integer.parseInt(buildProperties["MIN_SDK_VERSION"])
-        targetSdkVersion Integer.parseInt(buildProperties["SDK_VERSION"])
-        versionCode = libProperties['VERSION_CODE']
-        versionName libProperties['VERSION_NAME']
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
-}
-
-ext {
-    PUBLISH_ARTIFACT_ID = 'extensions-viewmodel'
-    POM_DESCRIPTION = 'Kotlin extensions for ViewModel from Android Architecture Components'
-}
-
-apply from: "${rootProject.projectDir}/gradle/mavencentral/publish.gradle"
