--- viewmodel/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest package="cz.ackee.extensions.viewmodel">
-
-    <application/>
-
-</manifest>
